python/pyproject.toml
# The Python interface must be built without build isolation (PEP517)
# due to its runtime and build time dependency on system built petsc4py and mpi4py.
# pip install -r build-requirements.txt
[build-system]
requires = [
      "scikit-build-core[pyproject]",
      "nanobind>=1.8.0",
      "petsc4py",
      "mpi4py",
]
build-backend = "scikit_build_core.build"

[project]
name = "dolfinx_contact"
version = "0.7.0.dev0"
description = 'Contact custom assemblers using DOLFINx and Basix'
readme = "README.md"
requires-python = ">=3.8.0"
license = { file = "../LICENSE" }
authors = [
      { email = "sr957@cam.ac.uk", name = "Sarah Roggendorf" },
      { email = "dokken@simula.no", name = "Jørgen S. Dokken" },
]
dependencies = [
      "numpy>=1.21",
      "cffi",
      "petsc4py",
      "mpi4py",
      "fenics-dolfinx>=0.8.0.dev0,<0.9.0",
]

[project.optional-dependencies]
lint = ["flake8", "mypy"]
test = ["pytest", "coverage"]
all = ["dolfinx_contact[lint,test]"]

[tool.scikit-build]
wheel.packages = ["dolfinx_contact"]
sdist.exclude = ["*.cpp"]
cmake.build-type = "Release"

[tool.pytest]
junit_family = "xunit2"

[tool.mypy]
ignore_missing_imports = true


[tool.isort]
line_length = 120
src_paths = ["demos", "asimov-contact", "tests"]
known_first_party = ["dolfinx_contact", "basix", "dolfinx", "ffcx", "ufl"]
known_third_party = ["gmsh", "numba", "numpy", "pytest", "pyvista"]
known_mpi = ["mpi4py", "petsc4py"]
sections = [
      "FUTURE",
      "STDLIB",
      "MPI",
      "THIRDPARTY",
      "FIRSTPARTY",
      "LOCALFOLDER",
]
